Refactor and reorganize data folder
