+++ app/src/main/java/com/example/volcanoseason3/data/forecastLinks/ForecastLinksRepository.kt
@@ -1,4 +1,4 @@
-package com.example.volcanoseason3.data.gallery
+package com.example.volcanoseason3.data.forecastLinks
 
 class ForecastLinksRepository(
     private val dao: ForecastLinkDao
