+++ app/src/main/java/com/example/volcanoseason3/ui/home/ForecastLinksViewModel.kt
@@ -4,12 +4,10 @@ import android.app.Application
 import androidx.lifecycle.AndroidViewModel
 import androidx.lifecycle.asLiveData
 import androidx.lifecycle.viewModelScope
-import com.example.volcanoseason3.data.gallery.AppDatabase
-import com.example.volcanoseason3.data.gallery.ForecastLink
-import com.example.volcanoseason3.data.gallery.ForecastLinksRepository
-import kotlinx.coroutines.Dispatchers
+import com.example.volcanoseason3.data.forecastLinks.AppDatabase
+import com.example.volcanoseason3.data.forecastLinks.ForecastLink
+import com.example.volcanoseason3.data.forecastLinks.ForecastLinksRepository
 import kotlinx.coroutines.launch
-import kotlinx.coroutines.withContext
 
 class ForecastLinksViewModel(application: Application) : AndroidViewModel(application) {
     private val repository = ForecastLinksRepository(
